semana07/dia3/pos_backend/requirements.txt
django==3.2
djangorestframework
cloudinary
python-decouple
django-cors-headers
djangorestframework-simplejwt
psycopg2